Unify search topic page hit components into one
The design changed and they are no longer much different.
