+++ site/owid.scss
@@ -75,7 +75,6 @@
 @import "./search/SearchWritingResults.scss";
 @import "./search/SearchFlatArticleHit.scss";
 @import "./search/SearchStackedArticleHit.scss";
-@import "./search/SearchLargeTopicPageHit.scss";
 @import "./search/SearchTopicPageHit.scss";
 @import "./search/SearchResultHeader.scss";
 @import "./search/SearchAsDraft.scss";
