--- site/search/SearchLargeTopicPageHit.scss
@@ -1,38 +1,0 @@
-.search-large-topic-page-hit {
-    display: block;
-    text-decoration: none;
-    background-color: $blue-5;
-    padding: 16px;
-
-    &:hover {
-        .search-large-topic-page-hit__title-text {
-            text-decoration: underline;
-        }
-    }
-}
-
-.search-large-topic-page-hit__tag {
-    @include h6-black-caps;
-    color: $blue-60;
-    margin-bottom: 4px;
-}
-
-.search-large-topic-page-hit__title {
-    @include h3-bold;
-    margin: 0;
-    color: $blue-90;
-    margin-bottom: 8px;
-}
-
-.search-large-topic-page-hit__excerpt {
-    @include body-3-regular;
-    color: $blue-60;
-}
-
-.search-large-topic-page-hit__excerpt-paragraph {
-    margin: 0 0 16px;
-
-    &:last-child {
-        margin-bottom: 0;
-    }
-}
